refactor(tagger): extract helpers to reduce file from 379 to 68 lines
Split tagger.js into focused modules:
- tagger-config.js (57 lines): Supabase client, audience loading
- tagger-schema.js (106 lines): Zod schema builder
- tagger-validation.js (111 lines): code validation, mutual exclusivity
- tagger-prompt.js (83 lines): prompt building, debug logging
- tagger.js (68 lines): main orchestration (was 379)

All functions are now under 30 lines.
